requirements.txt
Django
django-cors-headers
djangorestframework
psycopg2
xhtml2pdf